+++ ecies/keys/helper.py
@@ -1,20 +1,30 @@
 from coincurve import PublicKey
 from coincurve.utils import GROUP_ORDER_INT
+from Crypto.Protocol.DH import (
+    import_x25519_private_key,
+    import_x25519_public_key,
+    key_agreement,
+)
 from Crypto.Random import get_random_bytes
 
 from ..config import EllipticCurve
-from ..consts import ETH_PUBLIC_KEY_LENGTH
+from ..consts import ETH_PUBLIC_KEY_LENGTH, SECRET_KEY_SIZE
 
 
 def is_valid_secret(curve: EllipticCurve, secret: bytes) -> bool:
+    if len(secret) != SECRET_KEY_SIZE:
+        return False
     if curve == "secp256k1":
         return 0 < bytes_to_int(secret) < GROUP_ORDER_INT
-    raise NotImplementedError
+    elif curve == "x25519":
+        return True
+    else:
+        raise NotImplementedError
 
 
 def get_valid_secret(curve: EllipticCurve) -> bytes:
     while True:
-        key = get_random_bytes(32)
+        key = get_random_bytes(SECRET_KEY_SIZE)
         if is_valid_secret(curve, key):
             return key
 
@@ -24,7 +34,10 @@ def get_public_key(
 ) -> bytes:
     if curve == "secp256k1":
         return PublicKey.from_secret(secret).format(compressed)
-    raise NotImplementedError
+    elif curve == "x25519":
+        return import_x25519_private_key(secret).public_key().export_key(format="raw")
+    else:
+        raise NotImplementedError
 
 
 def get_shared_point(
@@ -32,7 +45,14 @@ def get_shared_point(
 ) -> bytes:
     if curve == "secp256k1":
         return PublicKey(pk).multiply(sk).format(compressed)
-    raise NotImplementedError
+    elif curve == "x25519":
+        return key_agreement(
+            kdf=lambda x: x,
+            static_priv=import_x25519_private_key(sk),
+            eph_pub=import_x25519_public_key(pk),
+        )
+    else:
+        raise NotImplementedError
 
 
 def convert_public_key(
@@ -41,7 +61,10 @@ def convert_public_key(
     if curve == "secp256k1":
         # handle 33/65/64 bytes
         return PublicKey(pad_eth_public_key(data)).format(compressed)
-    raise NotImplementedError
+    elif curve == "x25519":
+        return import_x25519_public_key(data).export_key(format="raw")
+    else:
+        raise NotImplementedError
 
 
 # private below